Pull numerical inverse compute into the pipeline
